# Conflicts:
#	.github/workflows/main-ci.yml
#	.github/workflows/nightly-test-and-build.yml
#	.github/workflows/release-test-and-build.yml
#	.github/workflows/reusable-e2e-tests.yml
#	.github/workflows/reusable-make-command-smoke.yml
#	.github/workflows/reusable-test-verification-suite.yml
#	apps/lina-plugins~linapro_main
#	openspec/specs/framework-ai-capability-namespace/spec.md
#	openspec/specs/framework-ai-multimodal-capabilities/spec.md
#	openspec/specs/framework-ai-text-capability/spec.md
#	openspec/specs/framework-capability-registry/spec.md
#	openspec/specs/linactl-build-tool-consolidation/spec.md
#	openspec/specs/linapro-ai-core-plugin/spec.md
#	openspec/specs/plugin-dependency-management/spec.md
#	openspec/specs/plugin-host-domain-capabilities/spec.md
#	openspec/specs/plugin-host-service-extension/spec.md
#	openspec/specs/plugin-package-boundary-governance/spec.md
#	openspec/specs/plugin-runtime-loading/spec.md
#	openspec/specs/tenant-platform-access-control/spec.md
#	openspec/specs/user-auth/spec.md

+++ apps/lina-core/api/config/v1/config_create.go
@@ -19,5 +19,5 @@ type CreateReq struct {
 
 // CreateRes is the config creation response.
 type CreateRes struct {
-	Id int `json:"id" dc:"Parameter ID" eg:"1"`
+	Id int64 `json:"id" dc:"Parameter ID" eg:"1"`
 }
